+++ errors/errors.go
@@ -21,39 +21,12 @@ func ThrowKeybindError(duplicate string) {
 	os.Exit(1)
 }
 
-/*
-Note: not fully implemented
-*/
 func CatchConfigReadError(err error, instances map[string]types.Instance) {
-	fmt.Println("Config file broken. Attempting to heal...")
+	fmt.Println("Config file broken. Read error.")
 	panic(err)
 }
 
-/*
-Three step config error handler
-Step one is attempting to heal the config file
-by removing irregularities
-Step two is resetting the config file
-Step three is throwing the error
-Note: not implemented fully
-*/
 func CatchConfigWriteError(err error, inst types.Instance) {
-	fmt.Println("Config file broken. Attempting to heal...")
-	/*
-		HealConfig()
-		err = WriteConfig(inst)
-		if err != nil {
-			fmt.Println("Failed. Resetting config...")
-			ResetConfig()
-			err := WriteConfig(inst)
-			if err != nil {
-				fmt.Println("Fatal error: ")
-				panic(err)
-			} else {
-				fmt.Println("Success! Configurations fully reset & instance written to config")
-			}
-		} else {
-			fmt.Println("Success! Instance written to config.")
-		}
-	*/
+	fmt.Println("Config file broken.")
+	panic(err)
 }