+++ utility/utilities.go
@@ -28,7 +28,6 @@ func ClearTerminal() {
 
 /*
 Pushes an item to the top of instance slice
-Also provides item with index
 */
 func PushInstance(instances []types.Instance, pushing types.Instance) []types.Instance {
 	instances = append(instances, pushing)